Extract env helpers for dependency reasons.
